+++ src/Fantomas.Core.Tests/DotIndexedGetTests.fs
@@ -88,7 +88,9 @@ let mySampleMethod() =
             MultilineBlockBracketsOnSameColumn = true
             AlignFunctionSignatureToIndentation = true
             AlternativeLongMemberDefinitions = true
-            MultiLineLambdaClosingNewline = true }
+            MultiLineLambdaClosingNewline = true
+            MaxDotGetExpressionWidth = 50
+            MaxInfixOperatorExpression = 50 }
     |> prepend newline
     |> should
         equal