+++ src/Fantomas.Core.Tests/ModuleTests.fs
@@ -65,7 +65,7 @@ open Fantomas.SourceTransformer
 let sortAndDedup by l =
     // comment2
     l |> Seq.distinctBy by |> Seq.sortBy by |> List.ofSeq"""
-        config
+        { config with MaxInfixOperatorExpression = 50 }
     |> prepend newline
     |> should
         equal