+++ src/Fantomas.Core.Tests/NoTrailingSpacesTests.fs
@@ -13,7 +13,12 @@ let ``should not confuse me with an extra space at end of line v2`` () =
     (fun () -> parseCmdLine args |> ignore)
     |> should throw typeof<Argu.ArguParseException>"""
 
-    formatSourceString false codeSnippet { config with MaxInfixOperatorExpression = 90 }
+    formatSourceString
+        false
+        codeSnippet
+        { config with
+            MaxInfixOperatorExpression = 90
+            MaxArrayOrListWidth = 40 }
     |> should
         equal
         """let ``should not extrude without positive distance`` () =